requirements.txt
Flask==2.1.3
PyJWT==2.1.0
Werkzeug==2.0.3
requests==2.26.0
redis==4.1.0